extract getNews / getPage into class
